Creating a base class.
